requirements.txt
eel
Pillow
numpy